Move manager and pane logic to separate traits
